846: remove unnecessary/separate out components in main-map
