+++ app/components/main-map.js
@@ -9,7 +9,6 @@ import { alias } from '@ember/object/computed';
 import bblDemux from '../utils/bbl-demux';
 import Geometric from '../mixins/geometric';
 import drawnFeatureLayers from '../layers/drawn-feature';
-import highlightedLotLayer from '../layers/highlighted-lot';
 import selectedLayers from '../layers/selected-lot';
 
 const selectedFillLayer = selectedLayers.fill;
@@ -63,27 +62,11 @@ export default class MainMap extends Component {
 
   drawnFeatureLayers = drawnFeatureLayers;
 
-  highlightedLotFeatures = [];
-
   highlightedLayerId = null;
-
-  highlightedLotLayer = highlightedLotLayer;
 
   @computed('layerGroupsObject')
   get mapConfig() {
     return this.layerGroups;
-  }
-
-  @computed('highlightedLotFeatures')
-  get highlightedLotSource() {
-    const features = this.get('highlightedLotFeatures');
-    return {
-      type: 'geojson',
-      data: {
-        type: 'FeatureCollection',
-        features,
-      },
-    };
   }
 
   @computed('bookmarks.[]')
@@ -145,31 +128,6 @@ export default class MainMap extends Component {
   selectedFillLayer = selectedFillLayer;
 
   selectedLineLayer = selectedLineLayer;
-
-  @action
-  adjustBuildingsLayer(visible) {
-    const map = this.get('mainMap.mapInstance');
-    if (visible) {
-      map.flyTo({ pitch: 45 });
-    } else {
-      map.flyTo({ pitch: 0 });
-    }
-  }
-
-  @action
-  locateMe() {
-    const geolocateButton = document.querySelectorAll('.mapboxgl-ctrl-geolocate')[0];
-
-    if (geolocateButton) {
-      geolocateButton.click();
-      this.set('findMeDismissed', true);
-    }
-  }
-
-  @action
-  dismissFindMe() {
-    this.set('findMeDismissed', true);
-  }
 
   @action
   handleMapLoad(map) {
