--- app/layers/bookmarked-lots.js
@@ -1,10 +1,0 @@
-export default {
-  id: 'bookmarked-lots',
-  type: 'fill',
-  source: 'pluto',
-  'source-layer': 'pluto',
-  paint: {
-    'fill-opacity': 0.2,
-    'fill-color': '#FF0000',
-  },
-};
