--- app/layers/highlighted-lot.js
@@ -1,9 +1,0 @@
-export default {
-  id: 'highlighted-lot',
-  type: 'fill',
-  source: 'hovered-feature',
-  paint: {
-    'fill-opacity': 0.4,
-    'fill-color': '#999999',
-  },
-};
