--- app/layers/point-layer.js
@@ -1,11 +1,0 @@
-export default {
-  id: 'cd-circle',
-  type: 'circle',
-  source: 'currentAddress',
-  paint: {
-    'circle-radius': 8,
-    'circle-color': '#00A1F9',
-    'circle-stroke-width': 2,
-    'circle-stroke-color': '#FFFFFF',
-  },
-};
