+++ app/templates/components/locate-me-mobile.hbs
@@ -1,0 +1,16 @@
+{{#unless findMeDismissed}}
+  <div class="find-me">
+    <button
+        onclick={{action 'locateMe'}}
+        class="button hide-for-medium hide-for-print no-margin"
+        data-test-button="locate-me">
+      Locate&nbsp;Me
+    </button>
+    <button
+      onclick={{action 'dismissFindMe'}}
+      class="button hide-for-medium hide-for-print no-margin"
+      data-test-button="dismiss-find-me">
+      &times;
+    </button>
+  </div>
+{{/unless}}
