+++ app/templates/components/main-map.hbs
@@ -29,7 +29,6 @@
     hoveredFeature=map.hoveredFeature
     layerGroups=this.layerGroups
     interactivity=interactivity
-    highlightedLineFeatureLayer=highlightedLotLayer
     onLayerClick=(action 'handleLayerClick')
     onLayerHighlight=(action 'handleLayerHighlight')
      as |layers|}}
@@ -62,12 +61,6 @@
     {{/mapbox-gl-source}}
   {{/if}}
 
-  {{#if mainMap.currentAddress}}
-    {{#map.source sourceId='currentAddress' options=mainMap.addressSource as |source|}}
-      {{source.layer layer=mainMap.pointLayer}}
-    {{/map.source}}
-  {{/if}}
-
   {{#map-measurement-tools
     map=map.instance
     draw=this.draw as |measurement|
@@ -98,12 +91,4 @@
   {{mapbox-gl-on 'data' (action 'mapLoading') eventSource=map.instance}}
 {{/mapbox/basic-map}}
 
-{{!--
-  TODO: Geolocate Feature -> Move into separate component
---}}
-{{#unless findMeDismissed}}
-  <div class="find-me">
-    <button {{action 'locateMe'}} class="button hide-for-medium hide-for-print no-margin">Locate&nbsp;Me</button>
-    <button {{action 'dismissFindMe'}} class="button hide-for-medium hide-for-print no-margin">&times;</button>
-  </div>
-{{/unless}}
+{{locate-me-mobile}}
